Test plugin with multiple JaCoCo versions
